# [wasm] Improve SIMD vector equality operator